CLDSRRV-546: add POST object api handler
